end_to_end_tests/golden-record-custom/pyproject.toml
[tool.poetry]
name = "custom-e2e"
version = "0.1.0"
description = "A client library for accessing My Test API"

authors = []

readme = "README.md"
packages = [
    {include = "custom_e2e"},
]
include = ["CHANGELOG.md", "custom_e2e/py.typed"]


[tool.poetry.dependencies]
python = "^3.8"
httpx = "^0.15.0"
attrs = "^20.1.0"
python-dateutil = "^2.8.1"

[tool.black]
line-length = 120
target_version = ['py38']
exclude = '''
(
  /(
    | \.git
    | \.venv
    | \.mypy_cache
  )/
)
'''

[tool.isort]
line_length = 120
multi_line_output = 3
include_trailing_comma = true

[build-system]
requires = ["poetry>=1.0"]
build-backend = "poetry.masonry.api"